Fix type conversion warning.

+++ lib/libutils/ext/include/util.h
@@ -134,8 +134,8 @@ static inline uint64_t reg_pair_to_64(uint32_t reg0, uint32_t reg1)
 static inline void reg_pair_from_64(uint64_t val, uint32_t *reg0,
 				    uint32_t *reg1)
 {
-	*reg0 = val >> 32;
-	*reg1 = val;
+	*reg0 = (uint32_t)(val >> 32);
+	*reg1 = (uint32_t)val;
 }
 #endif
